python_api/app.py
import os
from flask import Flask, request, jsonify
from dotenv import load_dotenv
from langchain_openai import ChatOpenAI
from langchain_core.messages import HumanMessage, SystemMessage

load_dotenv()

app = Flask(__name__)

@app.route('/api/chat', methods=['POST'])
def chat_handler():
    try:
        data = request.get_json()
        user_message = data['messages'][-1]['content']

        model = ChatOpenAI(model="gpt-4o-mini", api_key=os.getenv("OPENAI_API_KEY"))
        
        messages = [
            SystemMessage(content="Eres un asistente de IA útil."),
            HumanMessage(content=user_message),
        ]
        
        response = model.invoke(messages)
        
        return jsonify({"response": response.content})

    except Exception as e:
        print(f"Error: {e}")
        return jsonify({"error": str(e)}), 500

if __name__ == '__main__':
    app.run(debug=True, port=5001)