python_api/requirements.txt
# requirements.txt
Flask
python-dotenv
langchain
langchain-openai
langchain-google-genai
langchain-deepseek
supabase